+++ lib/minfraud/components/account.rb
@@ -6,6 +6,8 @@ module Minfraud
     #
     # @see https://dev.maxmind.com/minfraud/#Account_(/account)
     class Account < Base
+      include Minfraud::Validates
+
       # A unique user ID associated with the end-user in your system. If your
       # system allows the login name for the account to be changed, this should
       # not be the login name for the account, but rather should be an internal
@@ -26,6 +28,17 @@ module Minfraud
       def initialize(params = {})
         @user_id      = params[:user_id]
         @username_md5 = params[:username_md5]
+
+        validate
+      end
+
+      private
+
+      def validate
+        return if !Minfraud.enable_validation
+
+        validate_string('user_id', 255, @user_id)
+        validate_md5('username_md5', @username_md5)
       end
     end
   end